% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_utils.R
\name{check_pulse_programs}
\alias{check_pulse_programs}
\title{Check Pulse Programs}
\usage{
check_pulse_programs(spectra_path, pulse_prog_value, sample_names)
}
\arguments{
\item{spectra_path}{A character specifying the path to the directory containing the spectra files.}

\item{pulse_prog_value}{A character specifying the pulse program to check for.}

\item{sample_names}{A character vector containing the names of the samples.}
}
\value{
A character vector containing the pulse programs for each sample.
}
\description{
This function checks the pulse programs used in the NMR spectra data.
}
