% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.dens.R
\name{compare.dens}
\alias{compare.dens}
\title{A density plot and compare Function}
\usage{
compare.dens(v1, v2, legend = c("v1", "v2"), main = NULL, ...)
}
\arguments{
\item{v1}{, give the data frame, object, for first graph}

\item{v2, }{give the data frame, object, for second graph}

\item{legend, }{in format legend= c("v1","v2"): to show the legend of the graph}

\item{main, }{provide a heading, default: v1 v/s v2}
}
\description{
This function allows you plot and compare any two graphs.
Takes the margin according as the max density
}
\examples{
pred=rnorm(1000,0,1) obs=rnorm(1000,0.2,2) compare.dens(pred,obs, legend=c("predicted","observed"))
compare.dens()
}
\keyword{density,}
\keyword{legend,}
\keyword{main,}
