% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_utils.R
\name{create_selected_samples_vector}
\alias{create_selected_samples_vector}
\title{Create a vector of selected samples}
\usage{
create_selected_samples_vector(selected_sample_list, split = ";")
}
\arguments{
\item{selected_sample_list}{A character string containing the selected sample IDs separated by commas.}

\item{split}{A character specifying the delimiter used to split the selected sample list. Defaults to ";".}
}
\value{
A numeric vector containing the selected sample IDs.
}
\description{
This function creates a vector of selected samples from a comma-separated string.
}
