% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_utils.R
\name{draw_compared_plots}
\alias{draw_compared_plots}
\title{Draw Compared Plots}
\usage{
draw_compared_plots(
  df_before,
  df_after,
  time,
  sample_name,
  title_before = " FID with the Group Delay (real part - zoom)",
  title_after = " FID after Group Delay removal (real part - zoom)",
  zoom = TRUE,
  limit = 300
)
}
\arguments{
\item{df_before}{A data frame containing the FID data before group delay removal.}

\item{df_after}{A data frame containing the FID data after group delay removal.}

\item{time}{A numeric vector representing the time axis.}

\item{sample_name}{A character specifying the name of the sample.}

\item{title_before}{A character specifying the title for the plot before group delay removal. Defaults to "FID with the Group Delay (real part - zoom)".}

\item{title_after}{A character specifying the title for the plot after group delay removal. Defaults to "FID after Group Delay removal (real part - zoom)".}

\item{zoom}{A logical indicating whether to zoom the plot. Defaults to TRUE.}

\item{limit}{An integer specifying the limit for zooming the plot. Defaults to 300.}
}
\value{
A grid.arrange object displaying the compared plots.
}
\description{
This function draws and compares two plots of FID data before and after group delay removal.
}
