% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_utils.R
\name{draw_raw_FID}
\alias{draw_raw_FID}
\title{Draw Raw Free Induction Decay (FID) Plot}
\usage{
draw_raw_FID(
  df,
  time,
  sample_name,
  title = " raw FID (real part)",
  zoom = FALSE,
  limit = 300
)
}
\arguments{
\item{df}{A data frame containing the FID data.}

\item{time}{A numeric vector representing the time axis.}

\item{sample_name}{A character specifying the name of the sample.}

\item{title}{A character specifying the title of the plot. Defaults to "raw FID (real part)".}

\item{zoom}{A logical indicating whether to zoom the plot. Defaults to FALSE.}

\item{limit}{An integer specifying the limit for zooming the plot. Defaults to 300.}
}
\value{
A ggplot object displaying the raw FID plot.
}
\description{
This function draws a plot of the raw Free Induction Decay (FID) data.
}
