% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_utils.R
\name{draw_solvent_suppresed_plots}
\alias{draw_solvent_suppresed_plots}
\title{Draw Solvent Suppressed Plots}
\usage{
draw_solvent_suppresed_plots(
  df_before = FIDdata_GDC,
  df_after = FIDdata_SS,
  sample_name,
  time = time,
  limit = 4000,
  solvent_re = SolventRe
)
}
\arguments{
\item{df_before}{A data frame containing the FID data before solvent suppression.}

\item{df_after}{A data frame containing the FID data after solvent suppression.}

\item{sample_name}{A character specifying the name of the sample.}

\item{time}{A numeric vector representing the time axis.}

\item{limit}{An integer specifying the limit for zooming the plot. Defaults to 4000.}

\item{solvent_re}{A numeric vector representing the estimated solvent residuals.}
}
\value{
A grid.arrange object displaying the plots.
}
\description{
This function draws plots of FID data with solvent residuals before and after suppression.
}
