% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_utils.R
\name{filtering_FIDs_samples}
\alias{filtering_FIDs_samples}
\title{Filtering FIDs Samples}
\usage{
filtering_FIDs_samples(FIDs, selected_sample_list, rm_duplicated_names = FALSE)
}
\arguments{
\item{FIDs}{A data frame containing the FIDs data.}

\item{selected_sample_list}{A character vector containing the names of the selected samples.}

\item{rm_duplicated_names}{A logical indicating whether to remove duplicated sample names. Defaults to FALSE.}
}
\value{
A filtered data frame containing the FIDs data.
}
\description{
This function filters the FIDs (Free Induction Decay) data based on the selected sample list.
}
