% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_analysis_utils.R
\name{generate_kegg_pathway_enrichment}
\alias{generate_kegg_pathway_enrichment}
\title{Generate KEGG pathway enrichment analysis}
\usage{
generate_kegg_pathway_enrichment(
  group,
  enrich_approx = "normality",
  results_table_method = "diffusion",
  results_graph_method = "diffusion",
  vertex.label.cex = 0.5,
  nlimit = 250
)
}
\arguments{
\item{group}{A character vector containing the names of metabolites for which enrichment analysis will be performed.}

\item{enrich_approx}{A character specifying the approximation method used for enrichment analysis. Defaults to "normality".}

\item{results_table_method}{A character specifying the method used for generating results table. Defaults to "diffusion".}

\item{results_graph_method}{A character specifying the method used for generating the results graph. Defaults to "diffusion".}

\item{vertex.label.cex}{A numeric value specifying the label size for vertices in the results graph. Defaults to 0.5.}

\item{nlimit}{An integer specifying the maximum number of nodes in the results graph. Defaults to 250.}
}
\value{
A list containing the defined compounds, enriched compounds, results table, and visual representation of the results graph.
}
\description{
This function performs KEGG pathway enrichment analysis for a given group of metabolites using the FELLA package.
}
