% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_utils.R
\name{generate_stacked_pdf_plot}
\alias{generate_stacked_pdf_plot}
\title{Generate Stacked PDF Plot}
\usage{
generate_stacked_pdf_plot(
  data,
  directory = "results/figures/",
  file_name,
  ppm,
  width = 16,
  height = 10
)
}
\arguments{
\item{data}{A data frame containing the spectral matrix data.}

\item{directory}{A character specifying the directory for saving the plot. Defaults to "results/figures/".}

\item{file_name}{A character specifying the name of the PDF file.}

\item{ppm}{A numeric vector representing the ppm axis.}

\item{width}{A numeric value specifying the width of the PDF plot. Defaults to 16.}

\item{height}{A numeric value specifying the height of the PDF plot. Defaults to 10.}
}
\value{
A stacked PDF plot saved in the specified directory.
}
\description{
This function generates a stacked PDF plot.
}
