% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_analysis_utils.R
\name{igraph_to_visnetwork}
\alias{igraph_to_visnetwork}
\title{Convert an igraph object to a visNetwork object}
\usage{
igraph_to_visnetwork(g)
}
\arguments{
\item{g}{An igraph object.}
}
\value{
A visNetwork object suitable for visualization.
}
\description{
This function converts an igraph object to a visNetwork object for visualization.
}
