% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_analysis_utils.R
\docType{data}
\name{metabolite_kegg_ids_dict}
\alias{metabolite_kegg_ids_dict}
\alias{no_kegg_id}
\title{List of metabolite names along with their KEGG IDs}
\format{
An object of class \code{character} of length 1.
}
\usage{
metabolite_kegg_ids_dict
}
\description{
This list provides the names of metabolites along with their corresponding KEGG IDs for reference and analysis purposes.
}
\keyword{datasets}
