% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_utils.R
\name{plotSpecMatPeps}
\alias{plotSpecMatPeps}
\title{Plot Spectral Matrix for 1D NMR Data}
\usage{
plotSpecMatPeps(
  Peps_df,
  ppm,
  ppm_lim = c(min(ppm), max(ppm)),
  K = 0.67,
  pY = 1,
  dppm_max = 0.2 * (max(ppm_lim) - min(ppm_lim)),
  asym = 1,
  beta = 0,
  cols = NULL
)
}
\arguments{
\item{ppm}{A numeric vector representing the ppm axis.}

\item{ppm_lim}{A numeric vector specifying the limits of ppm to be plotted. Defaults to c(min(ppm), max(ppm)).}

\item{K}{A numeric value specifying a scaling factor. Defaults to 0.67.}

\item{pY}{A numeric value specifying the maximum intensity. Defaults to 1.}

\item{dppm_max}{A numeric value specifying the maximum deviation in ppm. Defaults to 0.2 * (max(ppm_lim) - min(ppm_lim)).}

\item{asym}{A numeric value specifying the asymmetry factor. Defaults to 1.}

\item{beta}{A numeric value specifying the beta coefficient. Defaults to 0.}

\item{cols}{A vector of colors for plotting. Defaults to NULL, which generates a rainbow palette.}

\item{specmat}{A matrix containing the spectral data.}
}
\value{
A plot of the spectral matrix for 1D NMR data.
}
\description{
This function plots the spectral matrix for 1D NMR data.
}
