% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictplot.R
\name{predictplot}
\alias{predictplot}
\title{A Predicted plot Function}
\usage{
predictplot(predicts, observed)
}
\arguments{
\item{predicts}{, give the prediced data frame, object}

\item{observed, }{give the actual observed test values (precisely)}
}
\description{
This function allows you plot the predicted and observed data
}
\examples{
predictplot()
}
\keyword{density}
