\name{r_utils-package}
\alias{r_utils-package}
\alias{r_utils}
\docType{package}
\title{
\packageTitle{r_utils}
}
\description{
\packageDescription{r_utils}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{r_utils}
\packageIndices{r_utils}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{r_utils}

Maintainer: \packageMaintainer{r_utils}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
