% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_utils.R
\name{return_samples_to_reverse}
\alias{return_samples_to_reverse}
\title{Return Samples to Reverse}
\usage{
return_samples_to_reverse(matrix_ZOPC, threshold = 50)
}
\arguments{
\item{matrix_ZOPC}{A matrix containing data where negative percentages are to be checked.}

\item{threshold}{A numeric value specifying the threshold percentage. Samples with negative percentages greater than this threshold will be returned. Defaults to 50.}
}
\value{
A character vector containing the names of samples with negative percentages exceeding the threshold.
}
\description{
This function identifies samples with negative percentages exceeding a specified threshold in a matrix and returns their names.
}
