% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_utils.R
\name{reverse_selected_samples_axis}
\alias{reverse_selected_samples_axis}
\title{Reverse Selected Samples Axis}
\usage{
reverse_selected_samples_axis(selected_samples_to_reverse, matrix_ZOPC)
}
\arguments{
\item{selected_samples_to_reverse}{A character vector containing the names of samples to be reversed.}

\item{matrix_ZOPC}{A matrix containing data where the selected samples' axis will be reversed.}
}
\value{
A matrix with the selected samples' axis reversed.
}
\description{
This function reverses the selected samples' axis by multiplying their values by -1 in a matrix.
}
