\name{rutils-package}
\alias{rutils-package}
\alias{rutils}
\docType{package}
\title{
\packageTitle{rutils}
}
\description{
\packageDescription{rutils}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{rutils}
\packageIndices{rutils}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{rutils}

Maintainer: \packageMaintainer{rutils}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
