% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_utils.R
\name{save_plots}
\alias{save_plots}
\title{Save plots as SVG files}
\usage{
save_plots(
  plots_list,
  directory = "results/figures",
  suffix = "_raw_plot.svg",
  plot_size = c(10, 8)
)
}
\arguments{
\item{plots_list}{A list containing ggplot objects to be saved as plots.}

\item{directory}{A character specifying the directory where the plots will be saved. Defaults to "results/figures".}

\item{suffix}{A character specifying the suffix to be appended to the filenames of the saved plots. Defaults to "_raw_plot.svg".}

\item{plot_size}{A numeric vector specifying the width and height of the plot in inches. Defaults to c(10, 8).}
}
\description{
This function saves plots from a list of ggplot objects as SVG files.
}
