% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_analysis_utils.R,
%   R/preprocessing_utils.R
\name{save_rds}
\alias{save_rds}
\title{Save RDS file}
\usage{
save_rds(object, filename, directory = "results/tables")

save_rds(object, filename, directory = "results/tables")
}
\arguments{
\item{object}{The R object to be saved.}

\item{filename}{A character specifying the filename for the RDS file.}

\item{directory}{A character specifying the directory where the RDS file will be saved. Defaults to "results/tables".}
}
\description{
This function saves an R object in RDS format.

This function saves an R object in RDS format.
}
