% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_utils.R
\name{spectra_peps_to_asics_format}
\alias{spectra_peps_to_asics_format}
\title{Spectra PepsNMR to ASICs Format}
\usage{
spectra_peps_to_asics_format(peps_spectra)
}
\arguments{
\item{peps_spectra}{A matrix containing spectra data in PEPs format.}
}
\value{
A data frame with the spectra transformed to ASICs format.
}
\description{
This function transforms spectra data from PepsNMR package to ASICs format.
}
