% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_analysis_utils.R
\name{translate_ASICS_to_KEGG_ids}
\alias{translate_ASICS_to_KEGG_ids}
\title{Translate ASICS metabolite names to KEGG IDs}
\usage{
translate_ASICS_to_KEGG_ids(metabo_list, met_dict = metabolite_kegg_ids_dict)
}
\arguments{
\item{metabo_list}{A character vector of ASICS metabolite names.}

\item{met_dict}{A named character vector containing ASICS metabolite names as names and their corresponding KEGG IDs as values. Defaults to \code{metabolite_kegg_ids_dict}.}
}
\value{
A character vector of KEGG IDs corresponding to the input ASICS metabolite names.
}
\description{
This function translates ASICS metabolite names to their corresponding KEGG IDs using a provided dictionary.
}
