% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_package_meta.R
\name{update_package_meta}
\alias{update_package_meta}
\title{Update package meta data}
\usage{
update_package_meta(fixtures)
}
\arguments{
\item{fixtures}{(str) a character vetor of the fixes and features to mention in the NEWS.md file. Example:
c("speeking up the \code{get_data()} function",
"fixing the \code{pull_data()} function",
"minor")}
}
\value{
called for the side effect of updating the package meta data in the DESCRIPTION file and inserting
release-related information in the NEWS.md
}
\description{
The function updates the version and the date of the DESCRIPTION file, insert the new version
and the features/fix messages inside the NEWS.md file, provide informative message.
}
\examples{
\dontrun{
new_features <- c(
 "speeding up the get_data() function",
 "fixing the pull_data()",
 "minor"
)

update_package_meta(fixtures = new_features)
}

}
